requirements.txt
python-documentcloud==3.3.7
requests
